+++ bluetoe/bindings/nrf51.hpp
@@ -1,8 +1,8 @@
 #ifndef BLUETOE_BINDINGS_NRF51_HPP
 #define BLUETOE_BINDINGS_NRF51_HPP
 
-#include <link_layer.hpp>
-#include <ll_data_pdu_buffer.hpp>
+#include <bluetoe/link_layer.hpp>
+#include <bluetoe/ll_data_pdu_buffer.hpp>
 #include <cstdint>
 
 extern "C" void RADIO_IRQHandler(void);
