+++ bluetoe/link_layer/connection_details.cpp
@@ -1,4 +1,4 @@
-#include <connection_details.hpp>
+#include <bluetoe/connection_details.hpp>
 
 namespace bluetoe {
 namespace link_layer {
