+++ bluetoe/link_layer/include/bluetoe/connection_callbacks.hpp
