+++ bluetoe/link_layer/include/bluetoe/connection_event_callback.hpp
