+++ bluetoe/link_layer/include/bluetoe/l2cap_signaling_channel.hpp
@@ -3,7 +3,7 @@
 
 #include <cstdlib>
 #include <cstdint>
-#include <ll_meta_types.hpp>
+#include "ll_meta_types.hpp"
 
 namespace bluetoe {
