+++ bluetoe/link_layer/include/bluetoe/notification_queue.hpp
