+++ tests/att/notification_tests.cpp
@@ -1,7 +1,7 @@
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
 
-#include <notification_queue.hpp>
+#include <bluetoe/notification_queue.hpp>
 #include <test_servers.hpp>
 
 BOOST_AUTO_TEST_SUITE( notifications_by_value )
