+++ tests/link_layer/address_tests.cpp
@@ -1,4 +1,4 @@
-#include <address.hpp>
+#include <bluetoe/address.hpp>
 
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
