+++ tests/link_layer/channel_map_tests.cpp
@@ -1,4 +1,4 @@
-#include <channel_map.hpp>
+#include <bluetoe/channel_map.hpp>
 
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
