+++ tests/link_layer/connection_callbacks_tests.cpp
@@ -1,7 +1,7 @@
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
 
-#include <link_layer.hpp>
+#include <bluetoe/link_layer.hpp>
 
 #include "connected.hpp"
