+++ tests/link_layer/delta_time_tests.cpp
@@ -1,4 +1,4 @@
-#include <delta_time.hpp>
+#include <bluetoe/delta_time.hpp>
 
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
