+++ tests/link_layer/ll_advertising_tests.cpp
@@ -4,7 +4,7 @@
 #include <boost/test/included/unit_test.hpp>
 #include <boost/mpl/list.hpp>
 
-#include <link_layer.hpp>
+#include <bluetoe/link_layer.hpp>
 #include <meta_tools.hpp>
 #include <server.hpp>
 #include "test_radio.hpp"
