+++ tests/link_layer/ll_data_pdu_buffer_tests.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <buffer_io.hpp>
-#include <ll_data_pdu_buffer.hpp>
+#include <bluetoe/ll_data_pdu_buffer.hpp>
 
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
