+++ tests/link_layer/notification_queue_tests.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include <notification_queue.hpp>
+#include <bluetoe/notification_queue.hpp>
 
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
