+++ tests/link_layer/white_list_tests.cpp
@@ -1,4 +1,4 @@
-#include <white_list.hpp>
+#include <bluetoe/white_list.hpp>
 
 #define BOOST_TEST_MODULE
 #include <boost/test/included/unit_test.hpp>
