+++ tests/test_tools/address_io.cpp
@@ -1,4 +1,4 @@
-#include <address.hpp>
+#include <bluetoe/address.hpp>
 #include <ostream>
 #include <iomanip>
 #include <boost/io/ios_state.hpp>
